+++ src/CMakeLists.txt
@@ -280,8 +280,12 @@ add_executable(${MMEX_EXE} WIN32 MACOSX_BUNDLE
     model/UsageModel.h
     model/_all.h
 
+    panel/AssetList.cpp
+    panel/AssetList.h
     panel/AssetPanel.cpp
     panel/AssetPanel.h
+    panel/BudgetList.cpp
+    panel/BudgetList.h
     panel/BudgetPanel.cpp
     panel/BudgetPanel.h
     panel/DashboardPanel.cpp
@@ -296,6 +300,8 @@ add_executable(${MMEX_EXE} WIN32 MACOSX_BUNDLE
     panel/JournalPanel.h
     panel/ReportPanel.cpp
     panel/ReportPanel.h
+    panel/SchedList.cpp
+    panel/SchedList.h
     panel/SchedPanel.cpp
     panel/SchedPanel.h
     panel/StockList.cpp
