+++ src/panel/BudgetPanel.h
@@ -20,95 +20,18 @@
 
 #pragma once
 
-#include "_PanelBase.h"
 #include "model/BudgetModel.h"
+#include "_PanelBase.h"
+#include "BudgetList.h"
 
 class wxListCtrl;
 class wxListEvent;
-class BudgetPanel;
-
-/* Custom ListCtrl class that implements virtual LC style */
-class BudgetList : public ListBase
-
-{
-    DECLARE_NO_COPY_CLASS(BudgetList)
-    wxDECLARE_EVENT_TABLE();
-
-public:
-    enum LIST_ID
-    {
-        LIST_ID_ICON = 0,
-        LIST_ID_CATEGORY,
-        LIST_ID_FREQUENCY,
-        LIST_ID_AMOUNT,
-        LIST_ID_ESTIMATED,
-        LIST_ID_ACTUAL,
-        LIST_ID_NOTES,
-        LIST_ID_size, // number of columns
-    };
-
-private:
-    static const std::vector<ListColumnInfo> LIST_INFO;
-
-public:
-    BudgetList(BudgetPanel* cp, wxWindow *parent, const wxWindowID id);
-
-public:
-    /* required overrides for virtual style list control */
-    virtual wxString OnGetItemText(long item, long col_nr) const;
-    virtual wxListItemAttr *OnGetItemAttr(long item) const;
-    virtual int OnGetItemImage(long item) const;
-
-    void OnListItemSelected(wxListEvent& event);
-    void OnListItemActivated(wxListEvent& event);
-    void OnMouseMove(wxMouseEvent& event);
-
-private:
-    wxSharedPtr<wxListItemAttr> attr3_; // style3
-    BudgetPanel* cp_;
-    long selectedIndex_ = -1;
-};
 
 class BudgetPanel : public PanelBase
 {
     wxDECLARE_EVENT_TABLE();
 
 public:
-    BudgetPanel(int64 budgetYearID
-        , wxWindow *parent
-        , wxWindowID winid = wxID_ANY
-        , const wxPoint& pos = wxDefaultPosition
-        , const wxSize& size = wxDefaultSize
-        , long style = wxTAB_TRAVERSAL | wxNO_BORDER
-        , const wxString& name = "BudgetPanel");
-    ~BudgetPanel();
-
-    /* updates the checking panel data */
-    void initVirtualListControl();
-
-    /* Getter for Virtual List Control */
-    wxString getItem(long item, int col_id);
-
-    void DisplayBudgetingDetails(int64 budgetYearID);
-    int64 GetBudgetYearID()
-    {
-        return budgetYearID_;
-    }
-    wxString GetCurrentView()
-    {
-        return currentView_;
-    }
-    int GetItemImage(long item) const;
-    void OnListItemActivated(int selectedIndex);
-    int64 GetTransID(long item)
-    {
-        return budget_[item].first;
-    }
-
-    void RefreshList();
-
-    wxString BuildPage() const { return m_lc->BuildPage(GetPanelTitle()); }
-
     enum EIcons
     {
         ICON_RECONCILLED,
@@ -116,20 +39,31 @@ public:
         ICON_FOLLOWUP
     };
 
+public:
+    static const wxString VIEW_ALL;
+    static const wxString VIEW_NON_ZERO;
+    static const wxString VIEW_PLANNED;
+    static const wxString VIEW_INCOME;
+    static const wxString VIEW_EXPENSE;
+    static const wxString VIEW_SUMM;
+
 private:
-    std::vector<std::pair<int64, int64> > budget_;
-    std::map<int64, std::pair<int, bool > > displayDetails_; //map categid to level of the category, whether category is visible, and whether any subtree is visible 
+    std::vector<std::pair<int64, int64>> budget_;
+    //map categid to level of the category, whether category is visible,
+    // and whether any subtree is visible 
+    std::map<int64, std::pair<int, bool>> displayDetails_;
     std::map<int64, std::pair<double, double> > budgetTotals_;
     std::map<int64, BudgetFreq> budgetPeriod_;
     std::map<int64, double> budgetAmt_;
     std::map<int64, wxString> budgetNotes_;
     std::map<int64, std::map<int,double> > categoryStats_;
     bool monthlyBudget_;
-    wxSharedPtr<BudgetList> m_lc;
     wxString currentView_;
     int64 budgetYearID_;
     wxString m_monthName;
     wxString m_budget_offset_date;
+
+    wxSharedPtr<BudgetList> m_lc;
     wxStaticText* budgetReportHeading_ = nullptr;
     wxStaticText* income_estimated_ = nullptr;
     wxStaticText* income_actual_ = nullptr;
@@ -139,11 +73,44 @@ private:
     wxStaticText* expenses_diff_ = nullptr;
     wxButton* m_bitmapTransFilter = nullptr;
 
-    bool Create(wxWindow *parent, wxWindowID winid
-        , const wxPoint& pos = wxDefaultPosition
-        , const wxSize& size = wxDefaultSize
-        , long style = wxTAB_TRAVERSAL | wxNO_BORDER
-        , const wxString& name = "BudgetPanel");
+public:
+    BudgetPanel(
+        int64 budgetYearID,
+        wxWindow *parent,
+        wxWindowID winid = wxID_ANY,
+        const wxPoint& pos = wxDefaultPosition,
+        const wxSize& size = wxDefaultSize,
+        long style = wxTAB_TRAVERSAL | wxNO_BORDER,
+        const wxString& name = "BudgetPanel"
+    );
+    ~BudgetPanel();
+
+    /* updates the checking panel data */
+    void initVirtualListControl();
+
+    /* Getter for Virtual List Control */
+    wxString getItem(long item, int col_id);
+
+    void DisplayBudgetingDetails(int64 budgetYearID);
+    int64 GetBudgetYearID() { return budgetYearID_; }
+    wxString GetCurrentView() { return currentView_; }
+    int GetItemImage(long item) const;
+    void OnListItemActivated(int selectedIndex);
+    int64 GetTransID(long item) { return budget_[item].first; }
+
+    void RefreshList();
+
+    wxString BuildPage() const { return m_lc->BuildPage(GetPanelTitle()); }
+
+private:
+    bool Create(
+        wxWindow *parent,
+        wxWindowID winid,
+        const wxPoint& pos = wxDefaultPosition,
+        const wxSize& size = wxDefaultSize,
+        long style = wxTAB_TRAVERSAL | wxNO_BORDER,
+        const wxString& name = "BudgetPanel"
+    );
 
     void CreateControls();
     void sortList();
@@ -156,4 +123,3 @@ private:
     void OnViewPopupSelected(wxCommandEvent& event);
     void OnMouseLeftDown(wxCommandEvent& event);
 };
-
